Add files via upload
There are a number of costs for a particular company like R&D Spend, Administration, and Marketing Spend with respect to the states. And we need to figure out which company where to invest the money will be more profitable as well as which one is the most correlated with the profit.